Add debug visualization using graphviz
When the --graphviz_file=FILE is given (regardless of -v), we output
a DOT description of the CFG enriched with input and output tuples at
each point.

The code currently writes raw DOT and is likely to break when we need to
output labels that are not escaped for DOT. In that case, it is probably
an idea to use a graphviz crate.

Example usage:

    $ cargo +nightly run --release -- --graphviz_file=graph.dot inputs/issue-47680/nll-facts/main
    $ dot -o graph.svg -Tsvg graph.dot

Fixes #12

Co-authored-by: Yati Sagade <mail@ysagade.nl>
Co-authored-by: Andrea Lattuada <andrea.lattuada@inf.ethz.ch>'