location insensitive: handle placeholders explicitly instead of a by-product of liveness
Also, number and comment the rules, because the expansions are very verbose and mistakes are easy to make here.

We simplified the rules by modeling the fact that placeholder origins are live at all points by ... materializing this as actual tuples in the `origin_live_on_entry` relation, even though liveness analysis does not computes that.

This was also done because soufflé provides an easy `;` alternative operator, while datafrog requires us to do the alternative expansions manually. (Also we were lazy)