Make Polonius report any found errors by default
It's a bit confusing that running Polonius on some Rust code that you
know has a borrow error doesn't seem to actually do anything - it gives
the impression that no analysis is performed by default.

Of course, analysis *is* performed by default, but errors aren't
reported unless you opt in (which begs the question of what exactly the
tool is meant to be doing when you run it!)

This changes the behaviour to print error tuples by default if any are
found, for a better experience. The old behaviour can be re-enabled by
using `--no-show-tuples` to suppress tuple output even if errors are
present.